src/test/java/org/example/NotificadorEmailTest.java
package org.example;

import static org.mockito.Mockito.*;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

@ExtendWith(MockitoExtension.class)
class NotificadorEmailTest {
    @Mock
    private EmailCliente emailClienteMock;

    @Test
    public void testNotificar(){
        NotificadorEmail notificador = new NotificadorEmail(emailClienteMock);
        notificador.notificar("ejemplo@test.com","Hola Mundo!");

        verify(emailClienteMock, times(1)).enviarCorreo("ejemplo@test.com","Hola Mundo!");
    }
    @Test
    public void testNotificarConDireccionVacia() {
        NotificadorEmail notificador = new NotificadorEmail(emailClienteMock);
        notificador.notificar("", "Mensaje");

        verify(emailClienteMock, times(0)).enviarCorreo(anyString(), anyString());
    }

    @Test
    public void testNotificarConMensajeNulo() {
        NotificadorEmail notificador = new NotificadorEmail(emailClienteMock);
        notificador.notificar("ejemplo@test.com", null);
        verify(emailClienteMock, times(1)).enviarCorreo("ejemplo@test.com","");
    }
}